Converted to new symmetric mass ratio convention. Also moved geodesic functions to their own file / package.
