+++ tests/test_utility.py
@@ -4,11 +4,13 @@ from few.trajectory.ode import KerrEccEqFlux
 from few.utils.constants import YRSID_SI
 from few.utils.utility import (
     ELQ_to_pex,
-    get_fundamental_frequencies,
-    get_kerr_geo_constants_of_motion,
     get_mu_at_t,
     get_p_at_t,
-    get_separatrix,
+)
+from few.utils.geodesic import (
+    get_fundamental_frequencies,
+    get_kerr_geo_constants_of_motion,
+    get_separatrix,   
 )
